+++ backend/corpora/api/v1/auth/keys.py
@@ -4,7 +4,7 @@ from backend.corpora.common.auth0_manager import auth0_management_session
 from backend.corpora.common.corpora_config import CorporaAuthConfig
 from backend.corpora.common.utils.api_key import generate
 from backend.corpora.common.utils.http_exceptions import NotFoundHTTPException
-from backend.corpora.lambdas.api.v1.authentication import get_userinfo
+from backend.corpora.api.v1.authentication import get_userinfo
 
 
 def get(user: str):
