+++ backend/corpora/api/v1/collection_id/upload.py
@@ -1,17 +1,17 @@
 import requests
 from flask import make_response, g
 
-from .....api_server.db import dbconnect
-from .....common.upload import upload
-from .....common.utils.dl_sources.url import MissingHeaderException, from_url
-from .....common.utils.exceptions import (
+from backend.corpora.api_server.db import dbconnect
+from backend.corpora.common.upload import upload
+from backend.corpora.common.utils.dl_sources.url import MissingHeaderException, from_url
+from backend.corpora.common.utils.exceptions import (
     MaxFileSizeExceededException,
     InvalidFileFormatException,
     NonExistentCollectionException,
     InvalidProcessingStateException,
     NonExistentDatasetException,
 )
-from .....common.utils.http_exceptions import (
+from backend.corpora.common.utils.http_exceptions import (
     ForbiddenHTTPException,
     InvalidParametersHTTPException,
     TooLargeHTTPException,
