+++ backend/corpora/api/v1/curation/collections/collection_id/datasets/upload_link.py
@@ -1,6 +1,6 @@
 from flask import make_response
 
-from backend.corpora.lambdas.api.v1.collection_id.upload import upload_from_link
+from backend.corpora.api.v1.collection_id.upload import upload_from_link
 
 
 def put(collection_id: str, body: dict, token_info: dict):
