+++ backend/corpora/api/v1/gene_sets.py
@@ -1,10 +1,10 @@
 from flask import make_response, jsonify, g
 
-from ....common.corpora_orm import CollectionVisibility
-from ....common.entities.geneset import Geneset
-from ....api_server.db import dbconnect
-from ....common.utils.authorization_checks import is_user_owner_or_allowed
-from ....common.utils.http_exceptions import (
+from backend.corpora.common.corpora_orm import CollectionVisibility
+from backend.corpora.common.entities.geneset import Geneset
+from backend.corpora.api_server.db import dbconnect
+from backend.corpora.common.utils.authorization_checks import is_user_owner_or_allowed
+from backend.corpora.common.utils.http_exceptions import (
     ForbiddenHTTPException,
 )
