+++ tests/unit/backend/corpora/api_server/base_api_test.py
@@ -89,7 +89,7 @@ class BaseAuthAPITest(BaseAPITest):
     def setUp(self):
         super().setUp()
         self.mock_assert_authorized_token = patch(
-            "backend.corpora.lambdas.api.v1.authentication.assert_authorized_token",
+            "backend.corpora.api.v1.authentication.assert_authorized_token",
             side_effect=mock_assert_authorized_token,
         )
         self.mock_assert_authorized_token.start()
